GLM statistics for knockoff

Computes the signed maximum statistic $$W_j = \max(Z_j, \tilde{Z}_j) \cdot \mathrm{sgn}(Z_j - \tilde{Z}_j),$$ where \(Z_j\) and \(\tilde{Z}_j\) are the maximum values of \(\lambda\) at which the jth variable and its knockoff, respectively, enter the generalized linear model.

stat.glmnet_lambdasmax(X, X_k, y, family = "gaussian", ...)

Arguments

X

n-by-p matrix of original variables.

X_k

n-by-p matrix of knockoff variables.

y

vector of length n, containing the response variables. Quantitative for family="gaussian", or family="poisson" (non-negative counts). For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For family="multinomial", can be a nc>=2 level factor, or a matrix with nc columns of counts or proportions. For either "binomial" or "multinomial", if y is presented as a vector, it will be coerced into a factor. For family="cox", y should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable, with '1' indicating death, and '0' indicating right censored. The function Surv() in package survival produces such a matrix. For family="mgaussian", y is a matrix of quantitative responses.

family

response type (see above).

...

additional arguments specific to glmnet (see Details).

Value

A vector of statistics \(W\) of length p.

Details

This function uses glmnet to compute the regularization path on a fine grid of \(\lambda\)'s.

The additional nlambda parameter can be used to control the granularity of the grid of \(\lambda\) values. The default value of nlambda is 500.

If the family is 'binomial' and a lambda sequence is not provided by the user, this function generates it on a log-linear scale before calling 'glmnet'.

For a complete list of the available additional arguments, see glmnet::glmnet().

Examples

# Synthetic Data
set.seed(2024)
p=200; n=100; k=15
mu = rep(0,p); Sigma = diag(p)
X = matrix(rnorm(n*p),n)
nonzero = 1:k
beta = 3.5 * (1:p %in% nonzero)
y = X %*% beta + rnorm(n)

# Knockoff Procedure
Xk = create.knockoff(X = X, type = 'shrink', num = 2)
res = knockoff.filter(X,y,Xk,statistic = stat.glmnet_lambdasmax)
res$shat
#>  [1]   2   3   4   6   7   9  10  11  12  13  65  90 124