The Knockoff Filter

This function runs the Knockoffs procedure from start to finish, selecting variables relevant for predicting the outcome of interest.

knockoff.filter(
  X,
  y,
  Xk = NULL,
  knockoffs = create.second_order,
  statistic = stat.glmnet_coefdiff,
  fdr = 0.1,
  offset = 1,
  verbose = FALSE,
  ...
)

Arguments

X

n-by-p matrix or data frame of predictors.

y

response vector of length n.

Xk

Knockoff variables. If not provided, they will be generated using the knockoffs function.

knockoffs

method used to construct knockoffs for the \(X\) variables. It must be a function taking a n-by-p matrix as input and returning a n-by-p matrix of knockoff variables. By default, approximate model-X Gaussian knockoffs are used.

statistic

statistics used to assess variable importance. By default, a lasso statistic with cross-validation is used. See the Details section for more information.

fdr

target false discovery rate (default: 0.1).

offset

either 0 or 1 (default: 1). This is the offset used to compute the rejection threshold on the statistics. The value 1 yields a slightly more conservative procedure ("knockoffs+") that controls the false discovery rate (FDR) according to the usual definition, while an offset of 0 controls a modified FDR.

verbose

Logical; if TRUE, prints progress messages during the generation of knockoff matrices. Default is FALSE.

Value

An object of class "knockoff.result". This object is a list containing at least the following components:

X

matrix of original variables

Xk

matrix of knockoff variables

statistic

computed test statistics

threshold

computed selection threshold

selected

named vector of selected variables

Details

This function creates the knockoffs, computes the importance statistics, and selects variables. It is the main entry point for the knockoff package.

The parameter `knockoffs` controls how knockoff variables are created. By default, the model-X scenario is assumed and a multivariate normal distribution is fitted to the original variables \(X\). The estimated mean vector and the covariance matrix are used to generate second-order approximate Gaussian knockoffs. In general, the function `knockoffs` should take a n-by-p matrix of observed variables \(X\) as input and return a n-by-p matrix of knockoffs. Two default functions for creating knockoffs are provided with this package.

In the model-X scenario, under the assumption that the rows of \(X\) are distributed as a multivariate Gaussian with known parameters, then the function `create.gaussian` can be used to generate Gaussian knockoffs, as shown in the examples below.

In the fixed-X scenario, one can create the knockoffs using the function `create.fixed`. This requires \(n \geq p\) and it assumes that the response \(Y\) follows a homoscedastic linear regression model.

For more information about creating knockoffs, type `??create`.

The default importance statistic is [stat.glmnet_coefdiff]. For a complete list of the statistics provided with this package, type `??stat`.

It is possible to provide custom functions for the knockoff constructions or the importance statistics. Some examples can be found in the vignette.

References

Candes et al., Panning for Gold: Model-free Knockoffs for High-dimensional Controlled Variable Selection, arXiv:1610.02351 (2016). [https://web.stanford.edu/group/candes/knockoffs/index.html](https://web.stanford.edu/group/candes/knockoffs/index.html)

Barber and Candes, Controlling the false discovery rate via knockoffs. Ann. Statist. 43 (2015), no. 5, 2055–2085.

Examples

# Linear Regression
set.seed(2024)
p=100; n=80; k=10; scale=3
Ac=1:k
rho=0.3; SigmaX <- toeplitz(rho^(0:(p-1)))
SigmaXhalf=chol(SigmaX)
beta=matrix(0,p,1)
beta[Ac]=sample(c(-1,1)*scale,k,replace=TRUE)
X = matrix(rnorm(n*p),n)%*%SigmaXhalf
y = X %*% beta + rnorm(n)
Xk = create.knockoff(X = X, type = 'shrink', num = 2)
res1 = knockoff.filter(X, y, Xk,statistic =stat.glmnet_coefdiff,offset = 1,fdr = 0.1)
#> Loading required package: Matrix
#> Loaded glmnet 4.1-8
print(res1$shat)
#>  [1]  1  2  3  4  5  6  7  8  9 10

# Logistics Regression
pis <- plogis(X %*% beta)
Y <- factor(rbinom(n, 1, pis))
res2 =  knockoff.filter(X,Y,Xk,statistic =stat.glmnet_coefdiff,family='binomial',offset = 0,fdr = 0.2)
print(res2$shat)
#> [1]  1  2  4  5  6  7  8 10 49